userapi_mysql/app/requirements.txt
gunicorn
Flask
mysql-connector
flask_swagger_ui==3.36.0
